requirements.txt
aiohttp==3.12.15
email-validator==2.3.0
fastapi==0.116.1
pydantic-settings==2.10.1
sqlalchemy==2.0.43
uvicorn==0.35.0
